+++ example/lib/default.dart
@@ -17,25 +17,31 @@ class _DefaultPageState extends State<DefaultPage> with SingleTickerProviderStat
     _controller = RubberAnimationController(
         vsync: this,
         halfBoundValue: AnimationControllerValue(percentage: 0.5),
+        lowerBoundValue: AnimationControllerValue(pixel: 200),
         duration: Duration(milliseconds: 200)
     );
     _controller.addStatusListener(_statusListener);
-
+    _controller.animationState.addListener(_stateListener);
     super.initState();
   }
 
   @override
   void dispose() {
     _controller.removeStatusListener(_statusListener);
+    _controller.animationState.removeListener(_stateListener);
     super.dispose();
   }
 
+  void _stateListener() {
+    print("state changed ${_controller.animationState.value}");
+  }
+
   void _statusListener(AnimationStatus status) {
-    print("changed State ${_controller.animationState}");
+    print("changed status ${_controller.status}");
   }
 
   void _expand() {
-    _controller.launchTo(AnimationState.expanded);
+    _controller.expand();
   }
 
   @override